#104 rename: Table 관련 모듈 파일 정리
