+++ fe/src/components/Table/TableBody/index.ts
@@ -1,0 +1,4 @@
+import TableBodyIssues from "./TableBodyIssues";
+import TableBodyLabels from "./TableBodyLabels";
+
+export { TableBodyIssues, TableBodyLabels };
