+++ fe/src/components/Table/TableHeader/index.ts
@@ -1,0 +1,4 @@
+import TableHeaderIssues from "./TableHeaderIssues";
+import TableHeaderLabels from "./TableHeaderLabels";
+
+export { TableHeaderIssues, TableHeaderLabels };
